Add ML Engine, AI Controller, full DB seeder & advanced features
Backend:
- MLController: 8 new ML endpoints (collaborative filter, price prediction,
  urgency engine, review intelligence, smart bundle, churn predictor,
  demand forecast, social proof)
- AIController: mood recs, dynamic pricing, smart seats, sentiment,
  personalized feed, NLP search, voice, agentic AI with tool calling
- DatabaseSeeder: 10 movies, 7 cities, 14 theatres, 168 shows, full seat/pricing
- Movie model with @Column(name="movie_cast") to avoid SQL reserved word
- SecurityConfig, WebConfig: CORS + Spring Security setup
- New repositories: Movie, City, Screen, SeatTypeShow, Theatre
- OpenAIService with 1h cache, AgentService, TMDBService

+++ src/main/java/Bookmyshow2/repositories/ShowRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ShowRepository extends JpaRepository<Show, Integer> {
+public interface ShowRepository extends JpaRepository<Show, Long> {
 
     @Override
     @NotNull
-    Optional<Show> findById(@NotNull Integer showId);
+    Optional<Show> findById(@NotNull Long showId);
 }
